src/test_parentnode.py
from parentnode import ParentNode
from leafnode import LeafNode
import unittest

class TestParentNode(unittest.TestCase):
    def test_node_tag(self):
        node = ParentNode(None, [LeafNode("b", "Bold text")])
        with self.assertRaises(ValueError):
            node.to_html()
    
    def test_node_children(self):
        node = ParentNode("b", None)
        node2 = ParentNode("b", [])
        with self.assertRaises(ValueError):
            node.to_html()
        
        with self.assertRaises(ValueError):
            node2.to_html()
    
    def test_to_html(self):
        node = ParentNode(
            "p",
            [
                LeafNode("b", "Bold text"),
                LeafNode(None, "Normal text"),
                LeafNode("i", "italic text"),
                LeafNode(None, "Normal text"),
            ],
        )

        self.assertEqual(node.to_html(), '<p><b>Bold text</b>Normal text<i>italic text</i>Normal text</p>')

if __name__ == '__main__':
    unittest.main()